Refactor codegen.ts
Splitting out the not exported code from codegen.ts to make it easier to read.
